+++ sketch/kolektor2/src/Average/Average.h
@@ -13,30 +13,30 @@ class Average {
   public:
     Average(Source * source);
     void setActual(float newValue);
-    unsigned short getErrors();
+    unsigned int getErrors();
     byte getWarnings();
     float getAverage();
     float getValue();
-    void takeReading(short newval);
+    void takeReading(int newval);
     void doReading();
   private:
     Source *source;
-    short values[7] = {0, 0, 0, 0, 0, 0, 0};
-    short avg = 0;
-    short actual = 0;
+    int values[7] = {0, 0, 0, 0, 0, 0, 0};
+    int avg = 0;
+    int actual = 0;
     boolean initialized = false;
     boolean error = false;
-    unsigned short errorsInRow = 0;
-    unsigned short index = 0;
+    unsigned int errorsInRow = 0;
+    unsigned int index = 0;
     byte warnings = 0;
     byte initSuccessLoopCount = 0;
-    short readSingle();
-    bool isInValidRange(short value);
+    int readSingle();
+    bool isInValidRange(int value);
     void addWarnReading();
     void addErrorReading();
     void startInitializing ();
-    void handleUnitialized(short newVal);
-    void handleNormal(short newVal);
+    void handleUnitialized(int newVal);
+    void handleNormal(int newVal);
 };
